+++ lib/saia_skills.dart
@@ -4,4 +4,5 @@ library;
 export 'src/saia_skill_catalog.dart';
 export 'src/saia_skill_codec.dart';
 export 'src/saia_skill_descriptor.dart';
+export 'src/saia_skill_frontmatter.dart';
 export 'src/saia_skill_polyglot.dart';